+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project.
 - `BotCommandScope` classes.
 - `scope` and `language_code` fields to `GetMyCommands` request class.
 - `scope` and `language_code` fields to `SetMyCommands` request class.
+- `DeleteMyCommands` request class.
 
 ### Fixed
 - typing for `Message` object.